requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
Pillow>=10.2.0
markdown>=3.5.1
pydub>=0.25.1
feedparser>=6.0.10
openai>=1.3.5
beautifulsoup4>=4.12.2
google-api-python-client>=2.108.0
spacy>=3.7.2
# You'll need to run: python -m spacy download en_core_web_lg